requirements.txt
aniso8601==8.0.0
astroid==2.3.3
autopep8==1.5.1
blinker==1.4
click==7.1.1
Flask==1.1.2
Flask-JWT-Extended==3.24.1
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.1
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
pathtools==0.1.2
pycodestyle==2.5.0
PyJWT==1.7.1
pylint==2.4.4
pytz==2019.3
simplejson==3.17.0
six==1.14.0
SQLAlchemy==1.3.16
typed-ast==1.4.1
watchdog==0.10.2
Werkzeug==1.0.1
wrapt==1.11.2
